API Reference

RxEnvironments.add! — Function

add!(first::AbstractEntity{T,S,E}, second; environment=false) where {T,S,E}

Adds second to first. If environment is true, the AbstractEntity created for second will be labeled as an environment. The Markov Blankets for both entities will be subscribed to each other.

Arguments

first::AbstractEntity{T,S,E}: The entity to which second will be added.

second: The entity to be added to first.

passive=false: A boolean indicating whether second should be instantiated as an active entity.

RxEnvironments.AbstractEntity — Type

AbstractEntity{T}

The AbstractEntity type supertypes all entities. It describes basic functionality all entities should have. It is assumed that every entity has a markov blanket, which has actuators and sensors. The AbstractEntity also has a field that describes whether or not the entity is terminated.

RxEnvironments.terminate! — Function

terminate!(entity::AbstractEntity)

Terminate an entity by setting its is_terminated flag to true and severing off all subscriptions to and from the entity.

Arguments

entity::AbstractEntity: The entity to terminate.

RxEnvironments.is_subscribed — Function

is_subscribed(subject::AbstractEntity, target::AbstractEntity)

Check if subject is subscribed to target.

Arguments

subject::AbstractEntity: The entity that may be subscribed to target.

target::AbstractEntity: The entity that may be subscribed to by subject.

Returns

true if subject is subscribed to target, false otherwise.

RxEnvironments.update! — Function

update!(e::AbstractEntity{T,ContinuousEntity,E}) where {T,E}

Update the state of the entity e based on its current state and the time elapsed since the last update. Acts as state transition function.

Arguments

e::AbstractEntity{T,ContinuousEntity,E}: The entity to update.

RxEnvironments.send! — Function

send!(recipient::AbstractEntity, emitter::AbstractEntity, action::Any)

Send an action from emitter to recipient.

RxEnvironments.receive! — Function

receive!(recipient::AbstractEntity, emitter::AbstractEntity, observation::Any)

Receive an observation from emitter and update the state of recipient accordingly.

See also: RxEnvironments.send!

RxEnvironments.emits — Function

emits(subject, listener, observation)

Determines if an entity of the type of subject should emit an observation to an entity of the type of listener when presented with an obervation of type observation. Users should implement this function for their own entity and observation types to handle the logic of when to emit observations to which entities. By default, this function returns true.

Rocket.unsubscribe! — Function

Unsubscribes receiver from emitter. Any data sent from emitter to receiver will not be received by receiver after this function is called.